chore: split integration tests
