+++ tests/integration/test_integration_sentiment_analysis.py
@@ -1,71 +1,11 @@
-# tests/test_model_integration.py
-
 import os
-import tempfile
-import pytest
 from pathlib import Path
-from tests.utils.utils import generate_heart_data, verify_prediction, cleanup_files
-import smolmodels as sm
+
 import pandas as pd
+import pytest
 
-
-@pytest.fixture(scope="session")
-def test_dir():
-    """Create a temporary directory for test files"""
-    with tempfile.TemporaryDirectory() as tmpdir:
-        yield Path(tmpdir)
-
-
-@pytest.fixture
-def sample_data():
-    """Generate synthetic heart disease data for testing"""
-    return generate_heart_data(n_samples=200)
-
-
-@pytest.fixture
-def input_schema():
-    """Define input schema - using simple types as expected by the model"""
-    return {
-        "age": int,
-        "gender": int,
-        "cp": int,
-        "trtbps": int,
-        "chol": int,
-        "fbs": int,
-        "restecg": int,
-        "thalachh": int,
-        "exng": int,
-        "oldpeak": float,
-        "slp": int,
-        "caa": int,
-        "thall": int,
-    }
-
-
-@pytest.fixture
-def output_schema():
-    """Define output schema - using simple types as expected by the model"""
-    return {"output": int}
-
-
-@pytest.fixture
-def test_input():
-    """Define a consistent test input for predictions"""
-    return {
-        "age": 61,
-        "gender": 1,
-        "cp": 3,
-        "trtbps": 145,
-        "chol": 233,
-        "fbs": 1,
-        "restecg": 0,
-        "thalachh": 150,
-        "exng": 0,
-        "oldpeak": 2.3,
-        "slp": 0,
-        "caa": 0,
-        "thall": 1,
-    }
+import smolmodels as sm
+from tests.utils.utils import verify_prediction, cleanup_files
 
 
 @pytest.fixture
@@ -113,24 +53,6 @@ def run_around_tests(model_dir):
     cleanup_files(model_dir)
 
 
-def test_model_with_data_and_schema(sample_data, input_schema, output_schema, test_input):
-    """Test case where user provides data, input and output schema"""
-    model = sm.Model(
-        intent="predict the probability of heart attack based on patient features",
-        input_schema=input_schema,
-        output_schema=output_schema,
-    )
-
-    model.build(
-        datasets=[sample_data],
-        provider="openai/gpt-4o-mini",
-        max_iterations=2,
-        timeout=3600,
-    )
-    prediction = model.predict(test_input)
-    verify_prediction(prediction, output_schema)
-
-
 def test_sentiment_classification(sentiment_data):
     """Test sentiment classification model end-to-end"""
     model = sm.Model(
@@ -139,17 +61,25 @@ def test_sentiment_classification(sentiment_data):
         output_schema={"sentiment": "str"},
     )
 
+    dataset = sm.DatasetGenerator(
+        description="A dataset of text reviews and their corresponding sentiment labels.",
+        provider="openai/gpt-4o-mini",
+        schema={"text": "str", "sentiment": "str"},
+        data=sentiment_data,
+    )
+    dataset.generate(10)
+
     # Build the model
     model.build(
-        datasets=[sentiment_data],
+        datasets=[dataset],
         max_iterations=3,  # Keep iterations low for testing
-        timeout=3600,
+        timeout=360,  # Keep timeout low for testing
     )
 
     # Test predictions
     test_cases = [
         {"input": {"text": "This product exceeded my expectations!"}, "expected_schema": {"sentiment": "str"}},
-        {"input": {"text": "Very disappointing experience"}, "expected_schema": {"sentiment": "str"}},
+        {"input": {"text": "Very disappointing experience, I hate it"}, "expected_schema": {"sentiment": "str"}},
     ]
 
     for test_case in test_cases:
